UserPatches imports PatchesList component
